Add files via upload
Scraping website data can be done easily using Power Query in Excel or by using Power BI. Both these tools have the ability to scrape tabular data from a website. The applications will search for a table within in the website and allows you to import it into the data model or Excel.